restructuring code to work within xrtpy, test should work but required changes to pytest.ini on my end
