+++ xrtpy/visualization_GUI/fov/__init__.py
@@ -1,0 +1,2 @@
+from xrtpy.visualization_GUI.fov import metadata_downloader
+
